+++ lib/clixon/clixon_yang.h
@@ -214,6 +214,7 @@ cg_var    *yang_cv_get(yang_stmt *ys);
 int        yang_cv_set(yang_stmt *ys, cg_var *cv);
 cvec      *yang_cvec_get(yang_stmt *ys);
 int        yang_cvec_set(yang_stmt *ys, cvec *cvv);
+cg_var    *yang_cvec_add(yang_stmt *ys, enum cv_type type, char *name);
 uint16_t   yang_flag_get(yang_stmt *ys, uint16_t flag);
 int        yang_flag_set(yang_stmt *ys, uint16_t flag);
 int        yang_flag_reset(yang_stmt *ys, uint16_t flag);