+++ lib/clixon/clixon_yang_parse_lib.h
@@ -52,6 +52,7 @@
  * Prototypes
  */
 yang_stmt *yang_parse_file(FILE *fp, const char *name, yang_stmt *ysp);
+int        yang_file_find_match(clicon_handle h, const char *module, const char *revision, cbuf *fbuf);
 yang_stmt *yang_parse_filename(const char *filename, yang_stmt  *ysp);
 int        yang_parse_post(clicon_handle h, yang_stmt *yspec, int modmin);
 int        yang_spec_parse_module(clicon_handle h, const char *module,