+++ lib/src/clixon_proto_client.c
@@ -198,6 +198,7 @@ clicon_rpc_msg(clicon_handle      h,
     int     s = -1;
     int     eof = 0;
 
+    clicon_debug(CLIXON_DBG_DETAIL, "%s", __FUNCTION__);
 #ifdef RPC_USERNAME_ASSERT
     assert(strstr(msg->op_body, "username")!=NULL); /* XXX */
 #endif
@@ -245,6 +246,7 @@ clicon_rpc_msg(clicon_handle      h,
     }
     retval = 0;
  done:
+    clicon_debug(CLIXON_DBG_DETAIL, "%s %d", __FUNCTION__, retval);
     if (retdata)
         free(retdata);
     if (xret)
@@ -961,6 +963,7 @@ clicon_rpc_get(clicon_handle   h,
     yang_stmt         *yspec;
     cvec              *nscd = NULL;
     
+    clicon_debug(CLIXON_DBG_DETAIL, "%s", __FUNCTION__);
     if (session_id_check(h, &session_id) < 0)
         goto done;
     if ((cb = cbuf_new()) == NULL){
@@ -1044,6 +1047,7 @@ clicon_rpc_get(clicon_handle   h,
     }
     retval = 0;
   done:
+    clicon_debug(CLIXON_DBG_DETAIL, "%s %d", __FUNCTION__, retval);
     if (nscd)
         cvec_free(nscd);
     if (cb)